requirement.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
networkx
osmnx
paho-mqtt
joblib
psycopg2-binary
python-dotenv
